requirements.txt
playwright>=1.40.0
firebase-admin>=6.4.0
python-dotenv>=1.0.1